@@ -1,52 +1,108 @@
 package main
 
 import (
-	"flag"
+	"context"
 	"log"
+	"log/slog"
 	"os"
 	"os/signal"
 	"syscall"
 
+	"github.com/na4ma4/meshtastic-mqtt-bin-to-json/internal/mainconfig"
 	"github.com/na4ma4/meshtastic-mqtt-bin-to-json/internal/relay"
+	"github.com/na4ma4/meshtastic-mqtt-bin-to-json/internal/store"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
+// rootCmd represents the base command when called without any subcommands.
+var rootCmd = &cobra.Command{
+	Use:   "meshtastic-mqtt-relay",
+	Short: "Meshtastic MQTT Relay",
+	Long:  `Meshtastic MQTT Relay is a tool for converting Meshtastic messages over MQTT from protobuf to JSON.`,
+	RunE:  mainCmd,
+	Args:  cobra.NoArgs,
+}
+
+func init() {
+	cobra.OnInitialize(mainconfig.ConfigInit)
+
+	rootCmd.PersistentFlags().BoolP("debug", "d", false, "Debug output")
+	_ = viper.BindPFlag("debug", rootCmd.PersistentFlags().Lookup("debug"))
+	_ = viper.BindEnv("debug", "DEBUG")
+
+	rootCmd.PersistentFlags().StringP("broker", "b", "tcp://localhost:1883", "Source MQTT broker URL")
+	_ = viper.BindPFlag("broker", rootCmd.PersistentFlags().Lookup("broker"))
+	_ = viper.BindEnv("broker", "MQTT_BROKER")
+
+	rootCmd.PersistentFlags().StringP("clientid", "c", "meshtastic-mqtt-relay", "MQTT client ID")
+	_ = viper.BindPFlag("clientid", rootCmd.PersistentFlags().Lookup("clientid"))
+	_ = viper.BindEnv("clientid", "MQTT_CLIENTID")
+
+	rootCmd.PersistentFlags().StringP("username", "u", "", "MQTT username (optional)")
+	_ = viper.BindPFlag("username", rootCmd.PersistentFlags().Lookup("username"))
+	_ = viper.BindEnv("username", "MQTT_USERNAME")
+
+	rootCmd.PersistentFlags().StringP("password", "p", "", "MQTT password (optional)")
+	_ = viper.BindPFlag("password", rootCmd.PersistentFlags().Lookup("password"))
+	_ = viper.BindEnv("password", "MQTT_PASSWORD")
+
+	rootCmd.PersistentFlags().StringP("topic", "t", "msh/ANZ/2/e/#", "MQTT topic to subscribe to")
+	_ = viper.BindPFlag("topic", rootCmd.PersistentFlags().Lookup("topic"))
+	_ = viper.BindEnv("topic", "MQTT_TOPIC")
+
+	rootCmd.PersistentFlags().BoolP("dry-run", "n", false, "Dry run mode (optional)")
+	_ = viper.BindPFlag("dry-run", rootCmd.PersistentFlags().Lookup("dry-run"))
+	_ = viper.BindEnv("dry-run", "MQTT_DRY_RUN")
+
+	rootCmd.PersistentFlags().StringP("output", "o", "", "Output store directory (optional)")
+	_ = viper.BindPFlag("output", rootCmd.PersistentFlags().Lookup("output"))
+	_ = viper.BindEnv("output", "OUTPUT_DIRECTORY")
+}
+
 func main() {
-	// Configuration flags
-	sourceBroker := flag.String("source-broker", "tcp://localhost:1883", "Source MQTT broker URL")
-	sourceClientID := flag.String("source-client-id", "meshtastic-relay-source", "Source MQTT client ID")
-	sourceTopic := flag.String("source-topic", "msh/+/json/+/+", "Source MQTT topic to subscribe to")
+	_ = rootCmd.Execute()
+}
 
-	destBroker := flag.String("dest-broker", "tcp://localhost:1884", "Destination MQTT broker URL")
-	destClientID := flag.String("dest-client-id", "meshtastic-relay-dest", "Destination MQTT client ID")
-	destTopic := flag.String("dest-topic", "meshtastic/json", "Destination MQTT topic prefix")
+func mainCmd(_ *cobra.Command, _ []string) error {
+	log.Printf("Starting Meshtastic MQTT Relay")
+	log.Printf("Source: %s (topic: %s)", viper.GetString("broker"), viper.GetString("topic"))
 
-	username := flag.String("username", "", "MQTT username (optional)")
-	password := flag.String("password", "", "MQTT password (optional)")
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
 
-	flag.Parse()
+	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelInfo}))
+	if viper.GetBool("debug") {
+		logger = slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelDebug}))
+		logger.Debug("Debug logging enabled")
+	}
 
 	config := relay.Config{
-		SourceBroker:   *sourceBroker,
-		SourceClientID: *sourceClientID,
-		SourceTopic:    *sourceTopic,
-		DestBroker:     *destBroker,
-		DestClientID:   *destClientID,
-		DestTopic:      *destTopic,
-		Username:       *username,
-		Password:       *password,
+		Broker:   viper.GetString("broker"),
+		ClientID: viper.GetString("clientid"),
+		Username: viper.GetString("username"),
+		Password: viper.GetString("password"),
+		Topic:    viper.GetString("topic"),
 	}
 
-	log.Printf("Starting Meshtastic MQTT Relay")
-	log.Printf("Source: %s (topic: %s)", config.SourceBroker, config.SourceTopic)
-	log.Printf("Destination: %s (topic: %s)", config.DestBroker, config.DestTopic)
-
-	relay, err := relay.NewRelay(config)
-	if err != nil {
-		log.Fatalf("Failed to create relay: %v", err)
+	if outputDir := viper.GetString("output"); outputDir != "" {
+		config.Store = store.NewJSONDirStore(outputDir)
+		log.Printf("Output directory set to: %s", outputDir)
 	}
 
-	if err := relay.Start(); err != nil {
-		log.Fatalf("Failed to start relay: %v", err)
+	var client *relay.Relay
+	{
+		var err error
+		client, err = relay.NewRelay(config, logger)
+		if err != nil {
+			log.Printf("Failed to create relay: %v", err)
+			return err
+		}
+	}
+
+	if err := client.Start(ctx); err != nil {
+		log.Printf("Failed to start relay: %v", err)
+		return err
 	}
 
 	// Wait for interrupt signal
@@ -55,5 +111,7 @@ func main() {
 	<-sigChan
 
 	log.Println("Shutting down...")
-	relay.Stop()
+	client.Stop(ctx)
+
+	return nil
 }
